Fix incorrect CPLDebug or CPLString::Printf arguments
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@13859 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/gtiff/geotiff.cpp
@@ -3539,7 +3539,7 @@ CPLErr GTiffDataset::OpenOffset( TIFF *hTIFFIn, toff_t nDirOffsetIn,
     {
         for (int i = 0; i < nBands; ++i)
             GetRasterBand(i+1)->SetMetadataItem( "NBITS", 
-                                                 CPLString().Printf( "%ld", nBitsPerSample ),
+                                                 CPLString().Printf( "%d", (int)nBitsPerSample ),
                                                  "IMAGE_STRUCTURE" );
     }
